[22644] Change the report of patients with no Wisepill reports in 48 hours to a report of all patients, ordered by last Wisepill report, with the least recent first.
Note that patients with NO Wisepill reports will sort to the end. I believe
study participants get a demo of their Wisepill device when they're starting
so all participants should have at least that report in the database. If that's
not the case, I can see if the sorting can be changed to put no reports first,
but right now it appears that will be non-trivial in Django.

Fixes case 22644.